Refactor HTML test suite boilerplate to external package.
