+++ haddock-test/haddock-test.cabal
@@ -1,0 +1,26 @@
+name:                 haddock-test
+version:              0.0.1
+synopsis:             Test utilities for Haddock
+license:              BSD3
+author:               Simon Marlow, David Waern
+maintainer:           Simon Hengel <sol@typeful.net>, Mateusz Kowalczyk <fuuzetsu@fuuzetsu.co.uk>
+homepage:             http://www.haskell.org/haddock/
+bug-reports:          https://github.com/haskell/haddock/issues
+copyright:            (c) Simon Marlow, David Waern
+category:             Documentation
+build-type:           Simple
+cabal-version:        >= 1.10
+stability:            experimental
+
+library
+  default-language: Haskell2010
+  hs-source-dirs:   src
+  build-depends:    base, directory, process, filepath, Cabal, xml, syb
+
+  exposed-modules:
+    Test.Haddock.Config
+    Test.Haddock.Process
+    Test.Haddock.Xhtml
+
+  other-modules:
+    Test.Haddock.Utils
