+++ haddock.cabal
@@ -127,7 +127,7 @@ test-suite html-test
   default-language: Haskell2010
   main-is:          run.hs
   hs-source-dirs:   html-test
-  build-depends:    base, directory, process, filepath, Cabal, xml, syb
+  build-depends:    base, directory, process, filepath, Cabal, xml, syb, haddock-test
 
 test-suite hypsrc-test
   type:             exitcode-stdio-1.0
